LogServ/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
bcrypt==4.0.1
python-jose==3.3.0
pymongo==4.10.1
python-dotenv==1.0.1
google-cloud-storage==2.19.0
python-multipart==0.0.20
httpx==0.28.1
passlib==1.7.4
gunicorn==23.0.0
ffmpeg-python==0.2.0